overall box

+++ src/main/java/nsu/graphics/thirdlab/object/Point3D.java
@@ -1,4 +1,4 @@
 package nsu.graphics.thirdlab.object;
 
-public record Point3D(int x, int y, int z) {
+public record Point3D(double x, double y, double z) {
 }
